pyproject.toml
[project]
name = "pintle-injector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ansys-fluent-core>=0.38.0",
    "gmsh>=4.15.1",
    "matplotlib>=3.10.8",
]
